#27 - refactor code for updated organization, update README
