+++ src/admin/templates/wp-handle-row-actions.php
